+++ src/main/java/com/volyVary/repository/FournitureRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.volyVary.model.Fourniture;
 
-public interface FournitureRepository extends JpaRepository<Fourniture, Long> {
+public interface FournitureRepository extends JpaRepository<Fourniture, Integer> {
 }
